+++ tests/test_decorators.py
@@ -75,8 +75,17 @@ async def test_upgrade_cluster_timeout(
             await handler._subhandler(
                 logger=logging.getLogger(__name__),
                 runtime=datetime.timedelta(seconds=runtime),
-                status={},
+                status={
+                    "subhandlerStartedAt": {
+                        "ActionSubhandler": {
+                            "started": int(datetime.datetime.utcnow().timestamp())
+                            - runtime,
+                            "ref": hash,
+                        }
+                    }
+                },
                 annotations={},
+                patch={},
             )
         await assert_wait_for(
             True,
@@ -103,7 +112,16 @@ async def test_upgrade_cluster_timeout(
         res = await handler._subhandler(
             logger=logging.getLogger(__name__),
             runtime=datetime.timedelta(seconds=runtime),
-            status={},
+            status={
+                "subhandlerStartedAt": {
+                    "ActionSubhandler": {
+                        "started": int(datetime.datetime.utcnow().timestamp())
+                        - runtime,
+                        "ref": hash,
+                    }
+                }
+            },
             annotations={},
+            patch={},
         )
         assert res["result"] == {"success": True}